+++ DependencyInjection/Configuration.php
@@ -32,7 +32,7 @@ class Configuration implements ConfigurationInterface
 
         $node
             ->scalarNode('is_inline_edition_on')->defaultFalse()->end()
-            ->scalarNode('use_streamed_response')->defaultTrue()->end()
+            ->scalarNode('use_streamed_response')->defaultFalse()->end()
             ->scalarNode('multisite')->isRequired()->end()
 
             ->arrayNode('ignore_route_patterns')